requirements.txt
gradio==4.44.0
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3